+++ src/main/scala/singleton/ops/impl/GeneralMacros.scala
@@ -644,7 +644,7 @@ trait GeneralMacros {
 
 
   ///////////////////////////////////////////////////////////////////////////////////////////
-  // Experimenting
+  // XTypeOf Experimental
   ///////////////////////////////////////////////////////////////////////////////////////////
   def materializeOpVal[F](implicit ev0: c.WeakTypeTag[F]): MaterializeOpAuxVal =
   new MaterializeOpAuxVal(weakTypeOf[F])
