+++ src/main/scala/singleton/ops/impl/OpMacros.scala
@@ -24,9 +24,8 @@ object OpMacro {
 }
 /*******************************************************************************************************/
 
-
 /********************************************************************************************************
-  * Experimenting
+  * XTypeOf Experimental
   *******************************************************************************************************/
 @ bundle
 object XTypeOf {
